rename embed button dir
