+++ packages/donate-button-beta/src/components/embed-button/index.tsx
